Authentification and authorization implemented using access and refresh token

+++ api/app.js
@@ -8,6 +8,8 @@ const bodyParser = require('body-parser');
 // Load in the mongoose models
 const { List, Task, User } = require('./db/models');
 
+const jwt = require('jsonwebtoken');
+
 // Load middleware
 app.use(bodyParser.json());
 
@@ -15,10 +17,78 @@ app.use(bodyParser.json());
 app.use(function(req, res, next) {
     res.header("Access-Control-Allow-Origin", "*");
     res.header("Access-Control-Allow-Methods", "GET, POST, HEAD, OPTIONS, PUT, PATCH, DELETE");
-    res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
+    res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept, x-access-token, x-refresh-token");
 
+    res.header(
+        'Access-Control-Expose-Headers',
+        'x-access-token, x-refresh-token' 
+    );
     next();
 });
+
+let authenticate = (req, res, next) => {
+    let token = req.header('x-access-token');
+
+    jwt.verify(token, User.getJWTSecret(), (err, decoded) => {
+        if (err) {
+            res.status(401).send(err);
+        } else {
+            req.user_id = decoded._id;
+            next();
+        }
+    });
+}
+
+// Verify Refresh Token Middleware (which will be verifying the session)
+let verifySession = (req, res, next) => {
+    // grab the refresh token from the request header
+    let refreshToken = req.header('x-refresh-token');
+
+    // grab the _id from the request header
+    let _id = req.header('_id');
+
+    User.findByIdAndToken(_id, refreshToken).then((user) => {
+        if (!user) {
+            // user couldn't be found
+            return Promise.reject({
+                'error': 'User not found. Make sure that the refresh token and user id are correct'
+            });
+        }
+
+
+        // if the code reaches here - the user was found
+        // therefore the refresh token exists in the database - but we still have to check if it has expired or not
+
+        req.user_id = user._id;
+        req.userObject = user;
+        req.refreshToken = refreshToken;
+
+        let isSessionValid = false;
+
+        user.sessions.forEach((session) => {
+            if (session.token === refreshToken) {
+                // check if the session has expired
+                if (User.hasRefreshTokenExpired(session.expiresAt) === false) {
+                    // refresh token has not expired
+                    isSessionValid = true;
+                }
+            }
+        });
+
+        if (isSessionValid) {
+            // the session is VALID - call next() to continue with processing this web request
+            next();
+        } else {
+            // the session is not valid
+            return Promise.reject({
+                'error': 'Refresh token has expired or the session is invalid'
+            })
+        }
+
+    }).catch((e) => {
+        res.status(401).send(e);
+    })
+}
 
 /* ROUTE HANDLERS */
 
@@ -28,10 +98,14 @@ app.use(function(req, res, next) {
  * GET /lists
  * Purpose: Get all lists
  */
-app.get('/lists', (req, res) => {
+app.get('/lists', authenticate, (req, res) => {
     // Return an array of all the lists
-    List.find({}).then((lists) => {
+    List.find({
+        _userId: req.user_id
+    }).then((lists) => {
         res.send(lists);
+    }).catch((e) => {
+        res.send(e);
     });
 });
 
@@ -39,13 +113,14 @@ app.get('/lists', (req, res) => {
  * POST /lists
  * Purpose: Create a list
  */
-app.post('/lists', (req, res) => {
+app.post('/lists', authenticate, (req, res) => {
     // Create a new list and return the new list document back to the user (which includes the id)
     // The list information (fields) will be passed in via the JSON request body
     let title = req.body.title;
 
     let newList = new List({
-        title
+        title,
+        _userId: req.user_id
     });
     newList.save().then((listDoc) => {
         // The full list document is returned
@@ -57,10 +132,11 @@ app.post('/lists', (req, res) => {
  * PATCH /lists/:id
  * Purpose: Update a specified list
  */
-app.patch('/lists/:id', (req, res) => {
+app.patch('/lists/:id', authenticate, (req, res) => {
     // update specified list with the new values in the json
-    List.findOneAndUpdate({ _id: req.params.id }, {
-        $set: req.body
+    List.findOneAndUpdate({ 
+        _id: req.params.id, 
+        _userId: req.user_id 
     }).then(() => {
         res.sendStatus(200);
     });
@@ -70,12 +146,16 @@ app.patch('/lists/:id', (req, res) => {
  * DELETE /lists/:id
  * Purpose: Delete a specified list
  */
-app.delete('/lists/:id', (req, res) => {
+app.delete('/lists/:id', authenticate, (req, res) => {
     // delete specified list
-    List.findOneAndRemove({ _id: req.params.id }, {
-
+    List.findOneAndRemove({
+        _id: req.params.id,
+        _userId: req.user_id
     }).then((removedListDoc) => {
         res.send(removedListDoc);
+
+        // delete all the tasks that are in the deleted list
+        deleteTasksFromList(removedListDoc._id);
     });
 });
 
@@ -83,7 +163,7 @@ app.delete('/lists/:id', (req, res) => {
  * GET /lists/:listId/tasks
  * Purpose: Get all tasks in a specific list
  */
-app.get('/lists/:listId/tasks', (req, res) => {
+app.get('/lists/:listId/tasks', authenticate, (req, res) => {
     // Return all tasks that belong to a specific list
     Task.find({
         _listId: req.params.listId
@@ -96,7 +176,7 @@ app.get('/lists/:listId/tasks', (req, res) => {
  * GET /lists/:listId/tasks/:taskId
  * Purpose: Get specific task by taskId
  */
-app.get('/lists/:listId/tasks/:taskId', (req, res) => {
+app.get('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
     // Return one task from list by specified taskId
     Task.findOne({ 
         _id: req.params.taskId,
@@ -110,14 +190,30 @@ app.get('/lists/:listId/tasks/:taskId', (req, res) => {
  * POST /lists/:listId/tasks
  * Purpose: Create task in specified list
  */
-app.post('/lists/:listId/tasks', (req, res) => {
-    // Create a new task in a list specified by listId
-    let newTask = new Task({
-        title: req.body.title,
-        _listId: req.params.listId
-    });
-    newTask.save().then((newTaskDoc) => {
-        res.send(newTaskDoc);
+app.post('/lists/:listId/tasks', authenticate, (req, res) => {
+
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            return true;
+        }
+
+        return false;
+    }).then((canCreateTask) => {
+        if (canCreateTask) {
+            // Create a new task in a list specified by listId
+            let newTask = new Task({
+                title: req.body.title,
+                _listId: req.params.listId
+            });
+            newTask.save().then((newTaskDoc) => {
+                res.send(newTaskDoc);
+            });
+        } else {
+            res.sendStatus(404);
+        }
     });
 });
 
@@ -125,15 +221,31 @@ app.post('/lists/:listId/tasks', (req, res) => {
  * PATCH lists/:listId/tasks/:taskId
  * Purpose: Update a task in specified list
  */
-app.patch('/lists/:listId/tasks/:taskId', (req, res) => {
-    // update specified task with the new values in the json
-    Task.findOneAndUpdate({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }, {
-        $set: req.body
-    }).then(() => {
-        res.send({message: 'Updated successfully!'})
+app.patch('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
+
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            return true;
+        }
+
+        return false;
+    }).then((canUpdateTask) => {
+        if (canUpdateTask) {
+            // update specified task with the new values in the json
+            Task.findOneAndUpdate({
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }, {
+                $set: req.body
+            }).then(() => {
+                res.send({message: 'Updated successfully!'})
+            });
+        } else {
+            res.sendStatus(404);
+        }
     });
 });
 
@@ -141,14 +253,29 @@ app.patch('/lists/:listId/tasks/:taskId', (req, res) => {
  * DELETE /lists/:listId/tasks/:taskId
  * Purpose: Delete a specified task in specific list
  */
-app.delete('/lists/:listId/tasks/:taskId', (req, res) => {
-    // Delete specified task in specific list
-    Task.findOneAndRemove({ 
-        _id: req.params.taskId, 
-        _listId: req.params.listId
-    }, {} ).then((removedTaskDoc) => {
-        res.send(removedTaskDoc);
-    });
+app.delete('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            return true;
+        }
+
+        return false;
+    }).then((canDeleteTask) => {
+        if (canDeleteTask) {
+            // Delete specified task in specific list
+            Task.findOneAndRemove({ 
+                _id: req.params.taskId, 
+                _listId: req.params.listId
+            }, {} ).then((removedTaskDoc) => {
+                res.send(removedTaskDoc);
+            });
+        } else {
+            res.sendStatus(404);
+        }
+    });    
 });
 
 /* USER ROUTES */ 
@@ -184,31 +311,55 @@ app.post('/users', (req, res) => {
  * POST /users/login
  * Purpose: Login
  */
+/**
+ * POST /users/login
+ * Purpose: Login
+ */
 app.post('/users/login', (req, res) => {
     let email = req.body.email;
     let password = req.body.password;
 
     User.findByCredentials(email, password).then((user) => {
         return user.createSession().then((refreshToken) => {
-            // Session created successfully - refreshToken returned
-            // now we generate an access auth token for the user
+            // Session created successfully - refreshToken returned.
+            // now we geneate an access auth token for the user
 
             return user.generateAccessAuthToken().then((accessToken) => {
-                // access auth token generated successfully, return an object containing the auth tokens
-                return {accessToken, refreshToken}
+                // access auth token generated successfully, now we return an object containing the auth tokens
+                return { accessToken, refreshToken }
             });
         }).then((authTokens) => {
-            res 
-                .header('x-refresh-token', authToken.refreshToken)
-                .header('x-access-token', authToken.accessToken)
-                .send(newUser);
-        }).catch((e) => {
-            res.status(400).send(e);
+            // Now we construct and send the response to the user with their auth tokens in the header and the user object in the body
+            res
+                .header('x-refresh-token', authTokens.refreshToken)
+                .header('x-access-token', authTokens.accessToken)
+                .send(user);
         })
-    })
+    }).catch((e) => {
+        res.status(400).send(e);
+    });
+})
+
+
+/**
+ * GET /users/me/access-token
+ * Purpose: Generates and return access token
+ */
+app.get('/users/me/access-token', verifySession, (req, res) => {
+    req.userObject.generateAccessAuthToken().then((accessToken) => {
+        res.header('x-access-token', accessToken).send({accessToken});
+    }).catch((e) => {
+        res.status(400).send(e);
+    });
 });
 
-
+let deleteTasksFromList = (_listId) => {
+    Task.deleteMany({
+        _listId
+    }).then(() => {
+        console.log('Tasks from ' + _listId + ' were deleted.');
+    });
+}
 app.listen(3000, () => {
     console.log("Server is listening on port 3000");
 });
